⚡️(medialive) switch to QVBR rate control mode
Since now we were using CBR (Constant BitRate) control mode. It is the
easiest one to implement. We decided to switch to QVBR, it's harder to
implement but we can save lot of bandwidth and also space without
loosing quality.
To understand QVBR and which value to use we follow this guide :
https://docs.aws.amazon.com/mediaconvert/latest/ug/cbr-vbr-qvbr.html