🚧(live) configure cloudfront distribution to stream live event
We would like to use cloudfront to distrubed live stream to our users.
For this we need new resources managed by terraform :
- new cloudfront distribution with default rule
- new policy to allow marsha user to update this policy

when a new live stack is created, the cloudfront distribution must be
updated with a new origin targeting the newly created mediapackage
channel endpoint and a new cache behavior to match this resource and use
the correct origin.

The main issue we have for now is a too long delay before the cloudfront
distribution is deployed.

TODO to complete this feature:

- use mediapackage endpoint while cloudfront is not deployed
- clean cloudfront distribution when the live stack is destroyed
- test the behaviour when a live is ongoing and the distribution is
updated. What happen ?