💄(frontend) pull margins out of <DashboardObjectProgress />
We want to reuse <DashboardObjectProgress /> in a new component
where we do not want margins. To enable this new use case, we can
remove the margin in the component and add it back outside the
component wherever we need it.

We also took this opportunity to rename it to reflect its wider
use than just in the LTI app dashboard.