✨(back) create and close a xmpp room MUC
To add more interactivity we want to add a chat when a live is played.
We choose to use the XMPP protocol for this. The back application is
responsible to create and close the room linked to the video. The room
is created when a live is started and closed when the live is ended.