🚚(frontend) clean built frontend files before each build
As webpack use hashes to name files it generate, old files stay in place. So
after several builds, we can have a js folder with a very large amount of files
which is not efficient and furthermore can produce issues and slow down script
executions.
In order to clean these files before each build, we need to put them in a
dedicated directory to not remove other js scripts placed in static/js directory